requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-api-python-client>=2.130.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pydantic>=2.8.0
python-dateutil>=2.9.0
requests>=2.32.0
apscheduler>=3.10.4
